feat: Add ArtificialEvolutionV2 class
The code changes in `evolution.py` add a new class `ArtificialEvolutionV2` that extends the `ArtificialEvolution` class. This new class introduces a modified `evaluate_fitness()` method and a `best()` method to return the best network from the population based on fitness. These changes enhance the functionality and flexibility of the evolutionary algorithm.